refactor: 답변 타이머 로직 커스텀 훅으로 분리
